feat: study, tab 타입 정의 (#35)
