+++ app/(default)/_component/Card.tsx
@@ -1,4 +1,4 @@
-type CardType = "gray" | "green";
+type CardType = "gray" | "green" | "none";
 type CardSizeType = "md" | "lg";
 
 export interface CardProps {
@@ -17,7 +17,11 @@ const Card = ({
   className,
 }: CardProps) => {
   const borderColor =
-    type === "green" ? " border-green-middle" : " border-gray-light";
+    type === "green"
+      ? " border-green-middle"
+      : type === "gray"
+      ? " border-gray-light"
+      : "border-none";
   const borderRadius = roundSize === "md" ? "rounded-md" : "rounded-[10px]";
   const borderShadow = isShadow ? "shadow-sm" : "none";
