+++ types/tab.d.ts
@@ -1,0 +1,9 @@
+export interface TabProps {
+  tabItems: TabItems[];
+  className?: string;
+}
+
+export interface TabItems {
+  name: string;
+  link: string;
+}
